+++ iGCS/UserInterface/Views/CompassView.m
@@ -84,7 +84,7 @@
     int startAng = ((int)_heading - 180)/5 * 5;
     const float startX = c.x + (startAng-_heading) * oneDegX;
     int ang = (startAng < 0) ? startAng += 360 : startAng;    
-    for (int i = 0; i < 360; i += 5, ang = (ang+5) % 360) {
+    for (NSUInteger i = 0; i < 360; i += 5, ang = (ang+5) % 360) {
         // Find the x position of this tick
         const float x = startX + i*oneDegX;
         if (x < -20 || x > w + 20) continue;
@@ -122,7 +122,7 @@
     //
     CGContextSaveGState(ctx);
     CGContextSetShadow (ctx, CGSizeMake (TICK_WIDTH, TICK_WIDTH), 2.5);
-    for (int i = 0; i < 2; i++) {
+    for (NSUInteger i = 0; i < 2; i++) {
         CGContextBeginPath(ctx);
         CGContextMoveToPoint(ctx,    c.x, c.y-h/2);
         CGContextAddLineToPoint(ctx, c.x, c.y+h/2);    
@@ -146,7 +146,7 @@
                         (__bridge id)[UIColor colorWithRed:0.1 green:0.1 blue:0.1 alpha:1.0].CGColor];
     CGFloat locations[] = {0.0, 1.0};
     CGGradientRef gradient = CGGradientCreateWithColors(colorSpace, (__bridge CFArrayRef) colors, locations);
-    for (int i = 0; i < 2; i++) {
+    for (NSUInteger i = 0; i < 2; i++) {
         CGRect subRect = CGRectMake(((i == 0) ? c.x - w/2 : c.x + w/6), c.y - h/2, w/3, h);
         CGPoint startPoint = CGPointMake(CGRectGetMinX(subRect), CGRectGetMinY(subRect));
         CGPoint endPoint = CGPointMake(CGRectGetMaxX(subRect), CGRectGetMinY(subRect));
